restructure the codebase
