+++ backend/app/services/storage/__init__.py
@@ -1,0 +1,2 @@
+from .redis_storage import RedisSessionStorage, session_storage
+from .storage import SessionStorage
